ssl_plugins: improve interaction between le and async tasks
Also improving the plugin management to use less implicit mask and more
explicit values. Currently we're supporting only the le plugin anyway.
The code still needs other improvements, on how we pass around the
consul_manager, for example. Tests are coming in a near future as well.

Plugins are now stored in a registry instead of automatically discovered
using directory navigation. They need to be explicitly registered, and
the function that register then also must be explicitly called.

Along with that, in order to reuse environment information that comes
with the task, we cannot call get_manager() inside the task, as the
environment variables are not available in the worker process. The
environment information may vary based on plan, so we save then in the
task, and then the work uses the saved information to build the instance
of the manager.