rpaas/consul_manager: use the provided config to connect to consul
That was my bad: tasks were not able to connect to Consul because they
were always loading Consul configuration (host, port and token) from the
environment, instead of the provided config. The worker has no
environment variables for communicating with Consul (or MongoDB, or
whatever). This information come from the task configuration.

It's available in the environment of the API, but not the worker.