ssl_plugins: add thread for renewing certificates
This thread will run in a given interval, adding a task to the queue,
and then the work will process the task, that then generate other tasks
for other workers to process.

The API process will run the thread. I still need to ensure some kind of
distributed lock to prevent multiple certificates renewing (we don't
want to fight LE rate limit).